infrastructure/onnx-runtime/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
onnxruntime==1.17.0
python-json-logger==2.0.7
aiofiles==23.2.1
